Drop executable shared lib support
In addition to not being portable, newer versions of Glibc (apparently
from v2.33) don't like this and refuse to load an 'executable lib'.
So I have to remove support.

In future commits I will add a build option to generate an Ag
executable, so we don't need to remove traditional Ag support.