~~ Relicensing project to Apache v2 ~~
Until now, libag was licensed under the MIT license. It turns out
that the ag project is licensed under the Apache v2 license. Using
mixed licenses in one project can lead to unwanted situations, so
this commit solves this problem by using the same license for
everything.

It is also important to point out that, before this commit, the
project continues to use the mixed licenses, and users are free
to choose to use this way, as long as they respect the limits of
both. In this scenario, users will automatically restrict
themselves to new commits.

Personally speaking, unless there is a very specific reason for
this, I recommend that everyone use the project in its latest
version and with the current license, i.e., Apache v2.