Thread-safe implementatino of ag_search()
This commit adds a 'thread-safe' version of ag_search(). Note that the
implementation is as rudimentary as possible and serializes all calls
to ag_search(). Parallel calls to ag_search_ts() will not be faster.

Part of this limitation concerns how ag is structured. See the issues
on GitHub for more information.

Note, however, that ag_search()/ag_search_ts() stays parallel for
multiple paths in a single call, so I would recommend using a single
call to multiples files instead of multiple calls whenever possible.