Add CMake Build for python bindings
Until now the build of Python bindings was restricted exclusively to
the Makefile. In order to make the code more portable and compile on
a larger number of systems this commit adds the build of python
bindings also in CMake.

Note that CMake is not my specialty and any help is very welcome to
make the build scripts better.